restructure code into generation, dataseting, preprocessing, learning_phase
