+++ README.md
@@ -1,0 +1,29 @@
+# Code Structure
+
+- [generation](generation)
+
+- [dataseting](dataseting)
+
+- [preprocessing](preprocessing)
+
+- [learning_phase](learning_phase)
+
+
+
+## generation
+
+- Generates data
+
+
+## dataseting
+
+- ?
+
+
+## preprocessing
+
+- converts raw data into clean format
+
+## learning_phase
+
+- deep learning module to predict y
